#vfb-135 - Refactors THREEDCanvas, code move to instance reducer. Work in progress; ThreeDCanvas class is now clean but need to finish the InstancesReducer to handle rest
