+++ applications/virtual-fly-brain/client/src/network/query.js
@@ -1,3 +1,5 @@
+import Resources from '@metacell/geppetto-meta-core/Resources';
+
 export const get_term_info = async (queryId) => {
   const url =`${API_URL}/get_term_info?id=${queryId}`;
   let response = await fetch(url)
@@ -33,4 +35,27 @@ export const get_instance = async (short_form) => {
   });
 
   return response;
+}
+
+export const get_3d_mesh = async (targetInstance) => {
+  let response = await fetch(targetInstance?.Images?.[Object.keys(targetInstance?.Images)[0]][0].obj)
+    .then(response => response.text())
+    .then(base64Content => {
+      const instance = {
+        "eClass": "SimpleInstance",
+        "id": targetInstance?.Id,
+        "name": targetInstance?.Name,
+        "type": { "eClass": "SimpleType" },
+        "visualValue": {
+          "eClass": Resources.OBJ,
+          'obj': base64Content
+        }, 
+        "visible" : true,
+        "color" : targetInstance?.color
+      }
+
+      return instance;
+    });
+
+    return response;
 }
