+++ applications/virtual-fly-brain/client/src/reducers/actions/types/getInstancesTypes.js
@@ -21,5 +21,6 @@ export const getInstancesTypes = Object.freeze({
   SHOW_LINES : 'SHOW_LINES',
   ADD_INSTANCE : "ADD_INSTANCE",
   ADD_SKELETON : "ADD_SKELETON",
-  LAUNCH_TEMPLATE : "LAUNCH_TEMPLATE"
+  LAUNCH_TEMPLATE : "LAUNCH_TEMPLATE",
+  UPDATE_INSTANCES : "UPDATE_INSTANCES",
 })
